replace old prototype with refactored version
